requirements.txt
mcp==1.24.0
sanic==25.3.0
click==8.3.1
aiohttp==3.13.2
aioresponses==0.7.8
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1
sanic-testing
uvloop; sys_platform != "win32"
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
PyYAML==6.0.3
redis==7.1.0
sqlalchemy[asyncio]
asyncpg
